requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.0
cron-descriptor==1.4.0
Django==4.2.4
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-filter==23.2
django-timezone-field==6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
flake8==6.1.0
idna==3.4
inflection==0.5.1
kombu==5.3.1
mccabe==0.7.0
packaging==23.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
